Decompilation improvements 03/27/2020
- Correctly support iterative structuring of loop regions with abnormal
exits (loop regions with abnormal entries are not supported yet).

- Add a few more boolean expression simplification routines.

- Unify the use of claripy ASTs. All claripy ASTs are removed at the
very end of the iterative structuring.

- Support continue statement.

- Refactor the code.
